Move vuex parallel implementation to test util folder
